server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.5.3
langchain==0.1.14
langsmith==0.1.17
langchain-community==0.0.31
langchain-core==0.1.43
langchain-openai==0.0.8
langchain-anthropic==0.1.9
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
qdrant-client==1.7.0
tavily-python==0.3.1
python-multipart==0.0.9
dspy==2.5.15
requests==2.31.0
bs4==0.0.2
beautifulsoup4==4.12.3
fastapi-cors==0.0.6
pyarrow==14.0.1
typer==0.9.0
pytz==2024.1
tenacity==8.2.3
sentence-transformers==2.2.2